# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "目前仅仅是一个demo",
        "title": "Rest API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/AddRole": {
            "post": {
                "description": "新增权限信息 请求主体: base64(str=aaaa) 成功输出null",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "新增权限信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "str",
                        "name": "str",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/DeleteRole": {
            "post": {
                "description": "删除权限信息 请求主体: base64(str=aaaa) 成功输出null",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "删除权限信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "str",
                        "name": "str",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetAllRoleForTransfer": {
            "post": {
                "description": "获取所有权限清单 请求主体: Null  成功输出[]TransferModel",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "获取所有权限清单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]TransferModel},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/GetAllRoleInfo": {
            "post": {
                "description": "前台条件获取权限信息 请求主体: base64(ParameterStr=aaaa\u0026PageSize=10\u0026CurrentPage=1) 成功输出[]User",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "前台条件获取权限信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ParameterStr",
                        "name": "ParameterStr",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PageSize",
                        "name": "PageSize",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "CurrentPage",
                        "name": "CurrentPage",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]User},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetDeptByUserIdForTransfer": {
            "post": {
                "description": "根据用户ID获取用户部门信息 请求主体: base64(userId=aaaa) 成功输出字符串",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dept"
                ],
                "summary": "根据用户ID获取用户部门信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":\"aaa,bbb,ccc\",\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetMenuByRoleIdForTree": {
            "post": {
                "description": "根据权限ID获取菜单树所有根结点 请求主体: base64(userId=aaaa) 成功输出字符串",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "根据权限ID获取菜单树所有根结点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":\"aaa,bbb,ccc\",\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/GetRoleByID": {
            "post": {
                "description": "根据权限ID获取权限信息 请求主体: base64(ID=aaaa) 成功输出Role",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "根据权限ID获取权限信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "ID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{Role},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/SetButtonByMenuIdRoleId": {
            "post": {
                "description": "根据菜单ID和角色ID设置按钮权限 请求主体: base64(menuId=aaaa\u0026menuId=aaaa) 成功输出User",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "根据菜单ID和角色ID设置按钮权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "菜单ID",
                        "name": "menuId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "权限ID",
                        "name": "roleId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "按钮ID，拼着接字符串",
                        "name": "menuStr",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/SetMenuRole": {
            "post": {
                "description": "设置权限菜单 请求主体: base64(roleId=aaaa) 成功输出User",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "设置权限菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "权限ID",
                        "name": "roleId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "菜单ID，拼着接字符串",
                        "name": "menuStr",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/UpdateRole": {
            "post": {
                "description": "修改权限信息 请求主体: base64(str=aaaa) 成功输出null",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "修改权限信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "str",
                        "name": "str",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}